requirements.txt
aws-cdk.assets==1.*
aws-cdk.aws-apigateway==1.*
aws-cdk.aws-apigatewayv2==1.*
aws-cdk.aws-apigatewayv2-integrations==1.*
aws-cdk.aws-applicationautoscaling==1.*
aws-cdk.aws-autoscaling==1.*
aws-cdk.aws-autoscaling-common==1.*
aws-cdk.aws-autoscaling-hooktargets==1.*
aws-cdk.aws-batch==1.*
aws-cdk.aws-certificatemanager==1.*
aws-cdk.aws-cloudformation==1.*
aws-cdk.aws-cloudfront==1.*
aws-cdk.aws-cloudwatch==1.*
aws-cdk.aws-codebuild==1.*
aws-cdk.aws-codecommit==1.*
aws-cdk.aws-codedeploy==1.*
aws-cdk.aws-codeguruprofiler==1.*
aws-cdk.aws-codepipeline==1.*
aws-cdk.aws-codepipeline-actions==1.*
aws-cdk.aws-cognito==1.*
aws-cdk.aws-dynamodb==1.*
aws-cdk.aws-ec2==1.*
aws-cdk.aws-ecr==1.*
aws-cdk.aws-ecr-assets==1.*
aws-cdk.aws-ecs==1.*
aws-cdk.aws-efs==1.*
aws-cdk.aws-elasticloadbalancing==1.*
aws-cdk.aws-elasticloadbalancingv2==1.*
aws-cdk.aws-events==1.*
aws-cdk.aws-events-targets==1.*
aws-cdk.aws-iam==1.*
aws-cdk.aws-kinesis==1.*
aws-cdk.aws-kinesisfirehose==1.*
aws-cdk.aws-kms==1.*
aws-cdk.aws-lambda==1.*
aws-cdk.aws-lambda-event-sources==1.*
aws-cdk.aws-logs==1.*
aws-cdk.aws-route53==1.*
aws-cdk.aws-route53-targets==1.*
aws-cdk.aws-s3==1.*
aws-cdk.aws-s3-assets==1.*
aws-cdk.aws-s3-deployment==1.*
aws-cdk.aws-sam==1.*
aws-cdk.aws-secretsmanager==1.*
aws-cdk.aws-servicecatalog==1.*
aws-cdk.aws-servicediscovery==1.*
aws-cdk.aws-sns==1.*
aws-cdk.aws-sns-subscriptions==1.*
aws-cdk.aws-sqs==1.*
aws-cdk.aws-ssm==1.*
aws-cdk.aws-stepfunctions==1.*
aws-cdk.cloud-assembly-schema==1.*
aws-cdk.core==1.*
aws-cdk.custom-resources==1.*
aws-cdk.cx-api==1.*
aws-cdk.pipelines==1.*
aws-cdk.region-info==1.*
requests==2.*
pytest
simplejson